+++ noodles-sam/src/alignment.rs
@@ -1,5 +1,6 @@
 //! Alignment record and fields.
 
+pub mod iter;
 pub mod record;
 
 pub use self::record::Record;