+++ noodles-sam/src/alignment/iter.rs
@@ -1,0 +1,5 @@
+//! Composable iterators for alignment records.
+
+mod pileup;
+
+pub use self::pileup::Pileup as Depth;